Yes, I realize this is a ridiculous premise.
I really used it as justification for the imagery on my webpage.

There was in the Matrix, a man born inside, who could shape it to his
will.

What if had been the only one with that unique if incidental genetic
code because the machines, try as they might, couldn't replicate human
genomes as uniformly or even randomly as they ever pleased.

What if substantiation could take place within the Matrix under certain
conditions, say -- a radioactive spider bite?

Yes, I'm laughing, too.